Replace Main with AutomatorView
